feat: Add refresh info box when new data arrives
