+++ src/css/_info-notice.scss
@@ -56,12 +56,10 @@
   box-shadow: inset 0 0 0 0.5px darken($color, 10%),
     inset 0 1px 0 0 rgba(255, 255, 255, 0.5), 0 0.5px 1px 0 rgba(0, 0, 0, 0.15);
 
-  outline: none;
   letter-spacing: 0.5px;
   &:active {
     background: darken($color, 8%);
     box-shadow: none;
-    outline: none;
   }
 }
