+++ src/js/components/TabSearch.js
@@ -1,34 +1,19 @@
 /* @flow */
-import {useDispatch, useSelector} from "react-redux"
-import React, {useEffect} from "react"
+import React from "react"
 
 import {DebugModal} from "./DebugModal"
 import {XDownloadProgress} from "./DownloadProgress"
 import ControlBar from "./ControlBar"
 import CurlModal from "./CurlModal"
 import EmptySpaceModal from "./EmptySpaceModal"
+import IngestRefresh from "./IngestRefresh"
 import IngestWarningsModal from "./IngestWarningsModal"
-import Search from "../state/Search"
 import SearchHeaderChart from "./SearchHeaderChart"
 import SearchResults from "./SearchResults/SearchResults"
-import Tab from "../state/Tab"
 import WhoisModal from "./WhoisModal"
 import ZQModal from "./ZQModal"
-import brim from "../brim"
-import submitSearch from "../flows/submitSearch"
 
 export default function TabSearch() {
-  let dispatch = useDispatch()
-  let firstVisit = !useSelector(Tab.currentEntry)
-  let space = useSelector(Tab.space)
-
-  useEffect(() => {
-    if (firstVisit) {
-      dispatch(Search.setSpanArgs(brim.space(space).everythingSpan()))
-      dispatch(submitSearch())
-    }
-  }, [firstVisit])
-
   return (
     <>
       <div className="search-page-header">
@@ -43,6 +28,7 @@ export default function TabSearch() {
       <ZQModal />
       <XDownloadProgress />
       <IngestWarningsModal />
+      <IngestRefresh />
     </>
   )
 }
